Refactor loader to use utils and helpers
