+++ .babelrc
@@ -1,0 +1,10 @@
+{
+  "presets": [[
+    "env",
+    {
+      "targets": {
+        "node": "4"
+      }
+    }
+  ]]
+}
