+++ package.json
@@ -2,15 +2,19 @@
   "name": "webpack-handlebars-loader",
   "version": "0.0.1",
   "description": "Handlebars loader for webpack-handlebars-static-file-generator",
-  "main": "src/index.js",
-  "devDependencies": {},
+  "main": "dist/index.js",
+  "devDependencies": {
+    "babel-cli": "^6.26.0",
+    "babel-preset-env": "^1.6.1"
+  },
   "dependencies": {
     "handlebars": "^4.0.11",
     "loader-utils": "^1.1.0",
     "lodash.merge": "^4.6.1"
   },
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "build": "babel src -d dist --copy-files"
   },
   "repository": {
     "type": "git",
